Shrink player some more
